requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-firestore==1.2.0
pillow==8.0.0